1.Delete old version Main form 2.Rename Main1 to Homepage 3.Delete redundant function
