+++ MathBrick/MathBrick/SignUp.cs
@@ -69,11 +69,6 @@ namespace MathBrick
             }
         }
 
-        private void SignUp_Load(object sender, EventArgs e)
-        {
-
-        }
-
         private void QuitButton_Click(object sender, EventArgs e)
         {
             this.Close();
