+++ MathBrick/MathBrick/SignUp.resx
@@ -120,44 +120,44 @@
   <assembly alias="System.Drawing" name="System.Drawing, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b03f5f7f11d50a3a" />
   <data name="skinPanel1.BackgroundImage" type="System.Drawing.Bitmap, System.Drawing" mimetype="application/x-microsoft.net.object.bytearray.base64">
     <value>
-        iVBORw0KGgoAAAANSUhEUgAAAIAAAACACAYAAADDPmHLAAAAAXNSR0IArs4c6QAAAARnQU1BAACxjwv8
-        YQUAAAAJcEhZcwAACxMAAAsTAQCanBgAAAhlSURBVHhe7Z19qC1VGcbvUfqACsUUP+4+s9Z8nHOM05fe
-        SlDzqhhGXAjEEAkM+icpwQIjE6mbXuriP6KoBP0jmUYqKPgVqJdQMbgKSkIaSH6Vce+FuhSmlmnPO/Pu
-        OWuvvWbPnLPP9jBrPz94OOe8612zzjvz7D1rZs+e2UYIIYQQQgghhBBCCCGEEEIIIYQQQgghhBBCCCGE
-        EEIIIWQzKbYXg9yYa6y1J2goXvJBXhTW7ob2Rq7dUquWPcbq6uoHsdFPL38m9tbc2Pcya7+rzXEChx+N
-        og9JsfMhcyjLsqO0/BFyk16teQfzJP2v/F4sLubaHCd4VexcWzlzImu/oeWPsGTtaWj/42h+en+eJDs0
-        JT5ggLNHC45fWZLt0vJDHIF18kqVZ94p+yT2t9oWHyEDYL93ZUzy65Oatfwx0jT9QpVnXltZXDmpMOY7
-        iH1Gm+MjYIDfaVM0SE1ujZMMcPL27R/HRr8PxrlYQ3FDA8w5NMCcQwPMOTTAnEMDrB85X5Al9kc4nDxT
-        Q/2FBlgf2WL2eSyjOj9g7Lu9NwENMJkswybOsmX9c1th0l95y9qtTf2EBmhGTgIh/92yn7X7M2svL4x9
-        211WmqRf0vR+QgM0I2cD3X4BvYi0I6rsnkIDNIPcJ91+Ab2UGfN1a+2HtUv/CBhAVtDPZi5jflkq1LbJ
-        CtTXzQCDvED+g1C1G2jWgTxJztBu/WI5y84LFBS1pGYtvxPLSZJhDvCn0LIc7dP0foH93GWBYqIWZvJX
-        afmdWB4MtqPf8NCvSQ9qer/APuyCQDFRC7P5S7X8TmBXddXoMsxzRZKs4l3hBvz9VzlCWDLmE5reL2R/
-        OFocVpCxf5ultnq8rnOAIXlifz7SH4eH2tR/AgaY+VGAN957Gp4ZUpM73noNgA1+Cvr9s+pvnls9bvWj
-        2tR/aIBuDAaDY+RqIbliWENxQAPMOTTAnEMDzDnTGGApTc+SQyD0eVI+JtVwK954NMBWMoUBFpD7+rBf
-        YezzGm9l2GcoDc8MjEEDNLFRA8gHIF6/t7SpFa8fDbCV0ABzDg0w53Q1gGxwqCZN0xWv39vaVNN00sTr
-        12SABWPMibqozoS++YsxaIAmuhhALnxE/B9eXhcdLIw5VRdT4+dpuGbHjh0fQFw+hx/J66h3Mu87/YjR
-        AE10MQBiD3k569GvdTE1fo6Ga/A/ne/nrFP/xmKOrJZGA0ykowHu8HI6C4eHt+hiavwcDdesfUN3w3od
-        i1molkYDTKSLAar75dgHoJcdld+hdySXTbntonuKojhOF1ODuNsvOAfITfp9tMlVOP4yW2SeQk07dTEl
-        iNMATXQxQAjMt3gUEAM0wJxDA8w5GzUA9u0f8vq9qU2teP1ogK1kowYQZMK11i99RMOtOGOV0vDMwBg0
-        QBPTGAC7gRNgguuyxP5U7q2j4Va88WiArWQaA2wUbzwaYCuhAeYc3wB4O9+Pt/aTZyl3PFEoZzMlNbnj
-        0QAO2Ief466c+ZA5R8sn2aK9NLyS4pXUrOWTJWsvDq2kmCU1a/n9Ry6awD7uB3mSXoI/649AuxKYBL4B
-        vTBjueOJQjmbKampHi+aOYCcjcuM/fOwMDke16bOBAzAo4C+kA2yT7mFwQzPaFNnaIAeg33ZZ0cKM/ZZ
-        beoMDdAz5KJHFPDVIkkvXDLpFW5h0EsSL9vS9Cyk11fFNLEJBpC7ZK3rTlneeG0GWEgX0y/qoVvjOPJU
-        j9zaL8v1hBqqkZrc8XprAPmKcp7YV91iJikz6U3atZFpDICNci7y/w4dkNunargVZ6xSGg5SJPb6Otfa
-        OwMbeAH/x7VryzMPS6xqqkA8DgNgBVzkFtJBrZ/RT2OAwpj7hv0yY27TcCvOWKU0HATtemOGSoVJ73ZM
-        gI2f7nHbRXJ3T20vQSwSAyzmn0QB/3OLmShr92vXRqY0QH3vHPz+bQ234oxVSsNBsIGf8PPFBNUj3sY3
-        PnRQjo60ewliEc0BjPkKirhDVoJ8Bu8WBh2WuAhv/zfLxZzarZFpDAAW0P983T+3zjeGeONNNkAO2xv7
-        ot8HNT7vx6DDoW8pIx6PAVzez6MAecUhdy829uPVq3KCrH0sS+yP0S14Ysobr20S2GgCT8GNL6CNBmii
-        qwGQ9z0vr1V4t/qmdh/Bz9PwRFpM0LjxBbTHaYDlk5aPRUH/qYtL7L3a1JmuBsCrWu6J5+a1Csv+iXYf
-        wc/TcBsy4bvR76t6NHT4NwTtcRpAwDH/1+QMIAp7wJ/9dqGzAaoJqBzyubmTdKDpsat+roYnEZztu5J5
-        T5MJ0B6vAaalqwEEK88ZTpIz5Muik4R5wukrKysf025jeOO1GaBp448dDTWZAG00QBPrMYCPfkX8UXm0
-        Kn43Gm7FG2+iAUJP/oQOw2i78DNwdGBv1641iNMATUxjAEzy7lrrZ36h4VacsUppOAja/d1OPeFrmhjK
-        zZ3LzgpiNEAT0xkgvWmtX7pHw604Y5XScBBMPv/g5I7N9sdNYP7l38oVcRqgiWkM8Onjj/8I+v8QG/8K
-        /+zbJLzxJhpA7sqtp5wfws9TNDxC+e1la3+DnH3YNYw9CwBxGqCJaQywUbzx2iaBUyM1uePRAA40QKTg
-        bfky6P62W5vTAJGCjf+0FItJ0+c0FCRgAHmixpWz1FaPF7UBMBm6SIv+ixRbzdTLSdrYrVqEkAFiV7QG
-        kI0cKliEooOPOEV8Zyg/ZknNWn50LOAd4FsocC9+Vs/KsfZWvAPsabo2QK4xxLHzIX8lxStzSE5pa/nx
-        0nUOIMgDEuUdQozzPul6bIh9Ivnda5uldkutWnbc6OVavw/dOpUQQgghhBBCCCGEEEIIIYQQQgghhBBC
-        CCGEEEIIIYQQQsjG2Lbt/24hb92uOwZHAAAAAElFTkSuQmCC
+        iVBORw0KGgoAAAANSUhEUgAAAIAAAACACAYAAADDPmHLAAAABGdBTUEAALGPC/xhBQAAAAlwSFlzAAAL
+        EwAACxMBAJqcGAAACGVJREFUeF7tnX2oLVUZxu9R+oAKxRQ/7j6z1nycc4zTl95KUPOqGEZcCMQQCQz6
+        JynBAiMTqZte6uI/oqgE/SOZRioo+BWol1AxuApKQhpIfpVx74W6FKaWac878+45a6+9Zs+cs8/2MGs/
+        P3g457zrXbPOO/PsPWtmz57ZRgghhBBCCCGEEEIIIYQQQgghhBBCCCGEEEIIIYQQQgghZDMptheD3Jhr
+        rLUnaChe8kFeFNbuhvZGrt1Sq5Y9xurq6gex0U8vfyb21tzY9zJrv6vNcQKHH42iD0mx8yFzKMuyo7T8
+        EXKTXq15B/Mk/a/8Xiwu5tocJ3hV7FxbOXMia7+h5Y+wZO1paP/jaH56f54kOzQlPmCAs0cLjl9Zku3S
+        8kMcgXXySpVn3in7JPa32hYfIQNgv3dlTPLrk5q1/DHSNP1ClWdeW1lcOakw5juIfUab4yNggN9pUzRI
+        TW6Nkwxw8vbtH8dGvw/GuVhDcUMDzDk0wJxDA8w5NMCcQwOsHzlfkCX2RzicPFND/YUGWB/ZYvZ5LKM6
+        P2Dsu703AQ0wmSzDJs6yZf1zW2HSX3nL2q1N/YQGaEZOAiH/3bKftfszay8vjH3bXVaapF/S9H5CAzQj
+        ZwPdfgG9iLQjquyeQgM0g9wn3X4BvZQZ83Vr7Ye1S/8IGEBW0M9mLmN+WSrUtskK1NfNAIO8QP6DULUb
+        aNaBPEnO0G79YjnLzgsUFLWkZi2/E8tJkmEO8KfQshzt0/R+gf3cZYFiohZm8ldp+Z1YHgy2o9/w0K9J
+        D2p6v8A+7IJAMVELs/lLtfxOYFd11egyzHNFkqziXeEG/P1XOUJYMuYTmt4vZH84WhxWkLF/m6W2eryu
+        c4AheWJ/PtIfh4fa1H8CBpj5UYA33nsanhlSkzveeg2ADX4K+v2z6m+eWz1u9aPa1H9ogG4MBoNj5Goh
+        uWJYQ3FAA8w5NMCcQwPMOdMYYClNz5JDIPR5Uj4m1XAr3ng0wFYyhQEWkPv6sF9h7PMab2XYZygNzwyM
+        QQM0sVEDyAcgXr+3tKkVrx8NsJXQAHMODTDndDWAbHCoJk3TFa/f29pU03TSxOvXZIAFY8yJuqjOhL75
+        izFogCa6GEAufET8H15eFx0sjDlVF1Pj52m4ZseOHR9AXD6HH8nrqHcy7zv9iNEATXQxAGIPeTnr0a91
+        MTV+joZr8D+d7+esU//GYo6slkYDTKSjAe7wcjoLh4e36GJq/BwN16x9Q3fDeh2LWaiWRgNMpIsBqvvl
+        2Aeglx2V36F3JJdNue2ie4qiOE4XU4O42y84B8hN+n20yVU4/jJbZJ5CTTt1MSWI0wBNdDFACMy3eBQQ
+        AzTAnEMDzDkbNQD27R/y+r2pTa14/WiArWSjBhBkwrXWL31Ew604Y5XS8MzAGDRAE9MYALuBE2CC67LE
+        /lTuraPhVrzxaICtZBoDbBRvPBpgK6EB5hzfAHg734+39pNnKXc8UShnMyU1uePRAA7Yh5/jrpz5kDlH
+        yyfZor00vJLildSs5ZMlay8OraSYJTVr+f1HLprAPu4HeZJegj/rj0C7EpgEvgG9MGO544lCOZspqake
+        L5o5gJyNy4z987AwOR7Xps4EDMCjgL6QDbJPuYXBDM9oU2dogB6DfdlnRwoz9llt6gwN0DPkokcU8NUi
+        SS9cMukVbmHQSxIv29L0LKTXV8U0sQkGkLtkretOWd54bQZYSBfTL+qhW+M48lSP3Novy/WEGqqRmtzx
+        emsA+YpynthX3WImKTPpTdq1kWkMgI1yLvL/Dh2Q26dquBVnrFIaDlIk9vo619o7Axt4Af/HtWvLMw9L
+        rGqqQDwOA2AFXOQW0kGtn9FPY4DCmPuG/TJjbtNwK85YpTQcBO16Y4ZKhUnvdkyAjZ/ucdtFcndPbS9B
+        LBIDLOafRAH/c4uZKGv3a9dGpjRAfe8c/P5tDbfijFVKw0GwgZ/w88UE1SPexjc+dFCOjrR7CWIRzQGM
+        +QqKuENWgnwG7xYGHZa4CG//N8vFnNqtkWkMABbQ/3zdP7fON4Z44002QA7bG/ui3wc1Pu/HoMOhbykj
+        Ho8BXN7PowB5xSF3Lzb249WrcoKsfSxL7I/RLXhiyhuvbRLYaAJPwY0voI0GaKKrAZD3PS+vVXi3+qZ2
+        H8HP0/BEWkzQuPEFtMdpgOWTlo9FQf+pi0vsvdrUma4GwKta7onn5rUKy/6Jdh/Bz9NwGzLhu9Hvq3o0
+        dPg3BO1xGkDAMf/X5AwgCnvAn/12obMBqgmoHPK5uZN0oOmxq36uhicRnO27knlPkwnQHq8BpqWrAQQr
+        zxlOkjPky6KThHnC6SsrKx/TbmN447UZoGnjjx0NNZkAbTRAE+sxgI9+RfxRebQqfjcabsUbb6IBQk/+
+        hA7DaLvwM3B0YG/XrjWI0wBNTGMATPLuWutnfqHhVpyxSmk4CNr93U494WuaGMrNncvOCmI0QBPTGSC9
+        aa1fukfDrThjldJwEEw+/+Dkjs32x01g/uXfyhVxGqCJaQzw6eOP/wj6/xAb/wr/7NskvPEmGkDuyq2n
+        nB/Cz1M0PEL57WVrf4Ocfdg1jD0LAHEaoIlpDLBRvPHaJoFTIzW549EADjRApOBt+TLo/rZbm9MAkYKN
+        /7QUi0nT5zQUJGAAeaLGlbPUVo8XtQEwGbpIi/6LFFvN1MtJ2titWoSQAWJXtAaQjRwqWISig484RXxn
+        KD9mSc1afnQs4B3gWyhwL35Wz8qx9la8A+xpujZArjHEsfMhfyXFK3NITmlr+fHSdQ4gyAMS5R1CjPM+
+        6XpsiH0i+d1rm6V2S61adtzo5Vq/D906lRBCCCGEEEIIIYQQQgghhBBCCCGEEEIIIYQQQgghhBBCyMbY
+        tu3/biFv3a47BkcAAAAASUVORK5CYII=
 </value>
   </data>
 </root>
